Output progress to UI app, and fix softlock that happened when user progresses through the app quickly.
Output wasn not being sent to the UI app because of the lack of using "2>&1" at the end of the line with dd to redirect the signal from dds own stdout to the stdout of update_backend.sh.

Progressing through the app quickly would often cause it to get stuck with the message "> Start" assigned to LINE.  This was because I had left an extra cout after Syscheck.